Rework the search/open API
search= will now search for files, case insensitively
open= will look for one file that matches the name, in any folder, and case insensitively.
If there are multiple matching files an error is shown, but not all those matching files
If no files match, there is a new option, `open_file_of_same_type`, which will make it look for other files with the same extension. If only one is found, then it will open it. This would be useful if you search for CounterfeitMonkey.gblorb but the zip has CounterfeitMonkey-v10.gblorb. But note that it could also return adventure.gblorb, if that's the only .gblorb in the zip.

Closes #15